+++ src/main/java/domain/info/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.info;
 
 public class Name {
