+++ src/main/java/domain/info/Rewards.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.info;
 
 import java.util.ArrayList;
 import java.util.List;
