+++ src/test/java/domain/RewardsTest.java
@@ -3,6 +3,8 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import domain.info.Names;
+import domain.info.Rewards;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
